alert-processor/requirements.txt
prometheus-api-client
pandas
numpy
requests  # For Slack/notification API calls
scikit-learn
pyyaml
joblib

